requirements.txt
Flask==2.3.3
slack_sdk
requests
redis
python-dotenv==1.0.0